Django/requirements.txt
Django==1.9.1
django-filter==0.13.0
django-registration-redux==1.3
django-smtp-ssl==1.0
djangorestframework==3.3.2
MySQL-python==1.2.5
wheel==0.24.0
python-memcached==1.57
